requirements.txt
beautifulsoup4==4.9.0
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9
csrf==0.1b1
cssselect2==0.2.2
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.0
dj-database-url==0.5.0
Django==2.2.5
django-allauth==0.41.0
django-appconf==1.0.3
django-bootstrap4==1.0.1
django-ckeditor==5.7.1
django-crispy-forms==1.7.2
django-extensions==2.2.9
django-import-export==2.0.2
django-js-asset==1.2.2
django-parler==2.0.1
django-paypal==1.0.0
djangorestframework==3.10.3
et-xmlfile==1.0.1
filelock==3.0.12
freeze==1.0.10
html5lib==1.0.1
idna==2.9
importlib-metadata==1.5.0
jdcal==1.4.1
MarkupPy==1.14
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
path==13.2.0
paypal==1.2.5
Pillow==7.0.0
pipenv==2018.11.26
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.1.0
Pyphen==0.9.5
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
soupsieve==2.0
sqlparse==0.3.0
tablib==1.1.0
tinycss2==1.0.2
urllib3==1.25.8
virtualenv==16.7.5
virtualenv-clone==0.5.3
WeasyPrint==51
webencodings==0.5.1
Werkzeug==1.0.1
xlrd==1.2.0
xlwt==1.3.0
zipp==3.0.0
